backend/requirements.txt
# Core Flask dependencies
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
gunicorn>=21.2.0

# Scientific computing (compatible with Python 3.13)
numpy>=1.26.2
scikit-learn>=1.4.0
pandas>=2.0.0
joblib>=1.3.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0

# API & Validation
marshmallow==3.20.1
flask-restx==1.3.0
jsonschema==4.20.0

# Security & Authentication
bcrypt==4.1.0
pyjwt==2.8.0
cryptography==42.0.0

# Monitoring & Logging
loguru==0.7.2

# Real-time & WebSocket
flask-socketio>=5.3.0
redis>=5.0.0
python-socketio>=5.8.0

# Data Export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# GeoIP and User Agent parsing
geoip2>=4.7.0
user-agents>=2.2.0

# Date & Time
python-dateutil==2.8.2

# HTTP Client (for external API calls)
requests==2.31.0

# Data Processing
scipy==1.12.0

# Development & Testing
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.0
flake8==7.0.0 